+++ src/Metadata/GraphQl/Query.php
@@ -12,21 +12,16 @@
 declare(strict_types=1);
 
 namespace ApiPlatform\Metadata\GraphQl;
-
 use ApiPlatform\State\OptionsInterface;
 
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 class Query extends Operation
 {
-    /**
-     * {@inheritdoc}
-     */
     public function __construct(
         ?string $resolver = null,
         ?array $args = null,
         ?array $links = null,
 
-        // abstract operation arguments
         ?string $shortName = null,
         ?string $class = null,
         ?bool $paginationEnabled = null,
@@ -39,7 +34,6 @@ class Query extends Operation
         ?bool $paginationClientPartial = null,
         ?bool $paginationFetchJoinCollection = null,
         ?bool $paginationUseOutputWalkers = null,
-        ?array $paginationViaCursor = null,
         ?array $order = null,
         ?string $description = null,
         ?array $normalizationContext = null,
@@ -54,11 +48,9 @@ class Query extends Operation
         ?string $deprecationReason = null,
         ?array $filters = null,
         ?array $validationContext = null,
-        mixed $input = null,
-        mixed $output = null,
-        /** @var array|bool|string|null $mercure */
+        $input = null,
+        $output = null,
         $mercure = null,
-        /** @var bool|string|null $messenger */
         $messenger = null,
         ?bool $elasticsearch = null,
         ?int $urlGenerationStrategy = null,
@@ -71,16 +63,63 @@ class Query extends Operation
         ?bool $forceEager = null,
         ?int $priority = null,
         ?string $name = null,
-        /** @var (callable(): mixed)|string|null $provider */
         $provider = null,
-        /** @var (callable(): mixed)|string|null $processor */
         $processor = null,
-        array $extraProperties = [],
         ?OptionsInterface $stateOptions = null,
+        array $extraProperties = [],
+
         protected ?bool $nested = null,
     ) {
-        parent::__construct(...\func_get_args());
-        $this->name = $name ?: 'item_query';
+        parent::__construct(
+            resolver: $resolver,
+            args: $args,
+            links: $links,
+            shortName: $shortName,
+            class: $class,
+            paginationEnabled: $paginationEnabled,
+            paginationType: $paginationType,
+            paginationItemsPerPage: $paginationItemsPerPage,
+            paginationMaximumItemsPerPage: $paginationMaximumItemsPerPage,
+            paginationPartial: $paginationPartial,
+            paginationClientEnabled: $paginationClientEnabled,
+            paginationClientItemsPerPage: $paginationClientItemsPerPage,
+            paginationClientPartial: $paginationClientPartial,
+            paginationFetchJoinCollection: $paginationFetchJoinCollection,
+            paginationUseOutputWalkers: $paginationUseOutputWalkers,
+            order: $order,
+            description: $description,
+            normalizationContext: $normalizationContext,
+            denormalizationContext: $denormalizationContext,
+            collectDenormalizationErrors: $collectDenormalizationErrors,
+            security: $security,
+            securityMessage: $securityMessage,
+            securityPostDenormalize: $securityPostDenormalize,
+            securityPostDenormalizeMessage: $securityPostDenormalizeMessage,
+            securityPostValidation: $securityPostValidation,
+            securityPostValidationMessage: $securityPostValidationMessage,
+            deprecationReason: $deprecationReason,
+            filters: $filters,
+            validationContext: $validationContext,
+            input: $input,
+            output: $output,
+            mercure: $mercure,
+            messenger: $messenger,
+            elasticsearch: $elasticsearch,
+            urlGenerationStrategy: $urlGenerationStrategy,
+            read: $read,
+            deserialize: $deserialize,
+            validate: $validate,
+            write: $write,
+            serialize: $serialize,
+            fetchPartial: $fetchPartial,
+            forceEager: $forceEager,
+            priority: $priority,
+            name: $name ?: 'item_query',
+            provider: $provider,
+            processor: $processor,
+            stateOptions: $stateOptions,
+            extraProperties: $extraProperties
+        );
     }
 
     public function getNested(): ?bool
@@ -96,3 +135,4 @@ class Query extends Operation
         return $self;
     }
 }
+