+++ src/Metadata/GraphQl/QueryCollection.php
@@ -19,15 +19,11 @@ use ApiPlatform\State\OptionsInterface;
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 final class QueryCollection extends Query implements CollectionOperationInterface
 {
-    /**
-     * {@inheritdoc}
-     */
     public function __construct(
         ?string $resolver = null,
         ?array $args = null,
         ?array $links = null,
 
-        // abstract operation arguments
         ?string $shortName = null,
         ?string $class = null,
         ?bool $paginationEnabled = null,
@@ -40,7 +36,6 @@ final class QueryCollection extends Query implements CollectionOperationInterfac
         ?bool $paginationClientPartial = null,
         ?bool $paginationFetchJoinCollection = null,
         ?bool $paginationUseOutputWalkers = null,
-        ?bool $paginationViaCursor = null,
         ?array $order = null,
         ?string $description = null,
         ?array $normalizationContext = null,
@@ -75,22 +70,56 @@ final class QueryCollection extends Query implements CollectionOperationInterfac
         protected ?OptionsInterface $stateOptions = null,
         array $extraProperties = [],
 
-        protected ?bool $nested = null,
+        ?bool $nested = null,
     ) {
-        parent::__construct(...\func_get_args());
-        $this->name = $name ?: 'collection_query';
-    }
-
-    public function getNested(): ?bool
-    {
-        return $this->nested;
-    }
-
-    public function withNested(?bool $nested = null): self
-    {
-        $self = clone $this;
-        $self->nested = $nested;
-
-        return $self;
+        parent::__construct(
+            resolver: $resolver,
+            args: $args,
+            links: $links,
+            shortName: $shortName,
+            class: $class,
+            paginationEnabled: $paginationEnabled,
+            paginationType: $paginationType,
+            paginationItemsPerPage: $paginationItemsPerPage,
+            paginationMaximumItemsPerPage: $paginationMaximumItemsPerPage,
+            paginationPartial: $paginationPartial,
+            paginationClientEnabled: $paginationClientEnabled,
+            paginationClientItemsPerPage: $paginationClientItemsPerPage,
+            paginationClientPartial: $paginationClientPartial,
+            paginationFetchJoinCollection: $paginationFetchJoinCollection,
+            paginationUseOutputWalkers: $paginationUseOutputWalkers,
+            order: $order,
+            description: $description,
+            normalizationContext: $normalizationContext,
+            denormalizationContext: $denormalizationContext,
+            security: $security,
+            securityMessage: $securityMessage,
+            securityPostDenormalize: $securityPostDenormalize,
+            securityPostDenormalizeMessage: $securityPostDenormalizeMessage,
+            securityPostValidation: $securityPostValidation,
+            securityPostValidationMessage: $securityPostValidationMessage,
+            deprecationReason: $deprecationReason,
+            filters: $filters,
+            validationContext: $validationContext,
+            input: $input,
+            output: $output,
+            mercure: $mercure,
+            messenger: $messenger,
+            elasticsearch: $elasticsearch,
+            urlGenerationStrategy: $urlGenerationStrategy,
+            read: $read,
+            deserialize: $deserialize,
+            validate: $validate,
+            write: $write,
+            serialize: $serialize,
+            fetchPartial: $fetchPartial,
+            forceEager: $forceEager,
+            priority: $priority,
+            name: $name ?: 'collection_query',
+            provider: $provider,
+            processor: $processor,
+            extraProperties: $extraProperties,
+            nested: $nested,
+        );
     }
 }